#54: Prevent fatal errors in 'factfinder/search_collection' resource model
For several possible reasons - server not reachable, return value not parsable, etc - the facade method _getFactFinderObject might return NULL instead of the requested object. Therefore, it is not correct to assume that the object will always be instantiated correctly, no matter what. Check if the object really exists before calling a method on it.